+++ core/terraform/airlock/locals.tf
@@ -5,6 +5,10 @@ locals {
   # STorage AirLock consolidated
   airlock_core_storage_name = lower(replace("stalairlock${var.tre_id}", "-", ""))
   
+  # Global Workspace Airlock Storage Account (Option B)
+  # STorage AirLock Global - all workspace stages for all workspaces
+  airlock_workspace_global_storage_name = lower(replace("stalairlockg${var.tre_id}", "-", ""))
+  
   # Container prefixes for stage segregation within consolidated storage account
   container_prefix_import_external    = "import-external"
   container_prefix_import_in_progress = "import-in-progress"